Create vfx_tutorial.md
Started blocking out a page for the VFX tutorial, setting up what to write and spending time in the VFX demo level to choose what sort of thing to be made. Also going to include a section that explains going through the demo level, as these should definitely go together.