Update vfx_tutorial.md
Added numbered steps to more parts of the post process section as it had started to become more of a reference/explanation.

Laid out loose ideas for how the interaction section should go. Spent time in Core setting up what that would be, and watching the heartbeast video to see what is done in that.